CLOUD-11977: Removed page service for setting titles
# Problem
The page titles in webadmin were incorrect on most pages in the application. This happened because the page service was overriding the page titles set by the page-meta service.

# Solution
Removed the pageService from the heath and monitoring tabs, and prevent the tite-resolver from setting the title to cloud in webadmin.

# Changes
- Adjusted the no cameras placeholder because the footer ran into the text.
![no_cameras_adjustment](/uploads/8afbae4892b06bdab6ac6683a066ac63/no_cameras_adjustment.png)