CLOUD-13787: Refactored the header to operate partially without an active system
# Problem
Channel partner users with service management cant view systems. As a result we have nothing to set with the header so its state is broken.

# Solution
Refactor the header so that it can operate on the services page without a system. The main change was introducing a new mode called channelPartnerServiceMode which happens when a user is trying to manage services for a system.

# How it works
1. The system guard checks if the user has a system. If they don't it checks the partner service.
2. Using info from the partner service the menus service is now set to the channel partner service mode. Here we create a L2 menu with only services as the option
3. User can do whatever they need for services.
4. When the user navigates away service mode is turned off.

(cherry picked from commit 7e3e9d26b25ec4b2afe06f29e1479b2c1c88c42d)