CLOUD-13793: Fixed issue where custom role didn't work in pre 6.0
# Problem
The handler for no role picked the last role in the accessRoles which can include custom roles. It wasn't a problem for offline cloud systems, but is for online ones.

# Solution
Replace the logic for taking the last role and replace it with the
customPermission role from the CONFIG

(cherry picked from commit 5cf0d11f5846c60a1e736a0904b92baa888a4f05)