CLOUD-13858: Updated check for removing users from orgs
# Problem
The check for updating the store was out of sync with the check for which api call we were using to remove a user from the org.

# Solution
Move the check to a variable outside the iif and subscribe and assign them to the same variable. This will prevent de-syncing between the api call and store update.

# Other Change
Updated the check to see if a user is being removed from all of their folders and treating it like an org delete.

(cherry picked from commit a671e2619dccbe1a9f6db602f31ab40f88ad263d)