CLOUD-13858: Fixed checks for updating the store when user is deleted
# Problem
When deleting a user from a folder they don't have groupRoles so an exception is thrown when checking which delete method we're going to use. Additionally, when deleting a user from a folder we need to update the currentGroup to remove them from the store.

# Solution
= Add optional chaining to account for when the user doesn't have group roles.
- When deleting a group user from a folder treat it like deleting a user from the org when updating the store.

(cherry picked from commit 5dccef0af5d5244c8f1988498d2df8ada3d86e37)