CLOUD-13909: Updated auth to check feature flags after login
# Problem
When the channelPartners flag is set to unknown the feature is false by default, and we never update the flag.

# Solution
Refactor the logic in the bind step of oauth to recheck the flag if channelPartners is off. Then if its on update the flag and carry on as usual.

(cherry picked from commit fd8507adfce22a5494aa734e5feb49e330f083eb)