CLOUD-13906: Force update docker image cache within runners on fresh build
# Changes
Force pull image whenever a build is triggered to ensure that pipeline has latest dependencies.

Remove pull_policy always on other jobs so that the image doesn't have to get pulled on every job.

Also add logging of commit that generated the build to confirm that correct image is being used within jobs.

![image](/uploads/310c127d083b3d64559e0ac9951563d7/image.png)
![image](/uploads/e0b8d85cef5650c0d3a5331d9eee3a62/image.png)

# Reason for the change
We use the same tag during the lifetime of the MR so it only gets pulled the first time it's built and cached is used therafter.