CLOUD-13676 CLOUD-13628 - Catch new Session Truncated error and redirect to OAuth
## Problem

There is a new session timeout option for Cloud sessions, and the error is called `sessionTruncated`. Each system can enforce a session time limit for cloud logins, so we need to catch these errors and force the user to relogin.

## Solution

We can use an interceptor and anytime we get a `sessionTruncated` error we redirect to OAuth. This in theory is very simple, but our OAuth logic is spread widely and duplicated throughout the app. This is just the first change that at least makes the functionality work.

I am still deciding whether I want to add on to the dispersed OAuth logic throughout the app or refactor things.

Major known issues:
* Because we do token exchanges on the home page with systems it will redirect to OAuth even on the home page. This is not good, but I think the main issue is that we shouldn't be doing token exchanges with every single system on the home page. So I opted to not catch this error at least for right now.
* If you are on the home page and you click into a system that requires reauth, then we redirect back to the home page. This is because I am just grabbing the current URL which is still the home page. Not sure the best approach to fixing this outside of manually trying to catch this in the interceptor. Since it's hard to even make this happen currently I will figure it out later
* I am clearing out all of the session storage and logging the user out when we redirect to OAuth. I feel like this was way overkill but I had too many bugs if I didn't do both. I will try to come back and see if there is a nicer way to do this, but I have to context switch for a week so I wanted a less buggy approach for now.