(releases_23.3.3->develop) CLOUD-13952 - Change millisecond enum names
## Problem

The millisecond enum (`MS`) that we use now is hard to quickly understand it's meaning when the surrounding context isn't perfect. Something like `MS.s` is hard to read

## Solution

We can still keep it short but use the full word of the time section. So changed to:

```ts
export enum MS {
    millisecond = 1,
    second = 1000,
    minute = MS.second * 60,
    hour = MS.minute * 60,
    day = MS.hour * 24,
}
```

(cherry-picked from commit fed7365caffb)