CLOUD-13966: Fix table row render delay
When `<nx-table>` initially loads, sometimes the rows render with a slight delay after the header, and sometimes not at all until the user interacts with the page (eg clicking or resizing).

The table rows are initialized by `initPageRows()` which was called in `ngOnChanges()` and `onResize()`. `initPageRows()` also only executes if the `#tableBodyContainer` element is available in the component view. Since ngOnChanges is run before view init, the rows were only being initialized in onResize which was causing the delay/flaky behavior.

This MR adds `initPageRows()` to an `ngAfterViewInit()` which seems to fix the issue