CLOUD-13964: Fixed table misalignment
# Problem
The columns could be shrunk to different sizes depending on its content. Since the header / body are not connected like a normal table they could theoretically diverge.

# Solution
Add a minmax for each column. In this case I took the ideal column size from figma to do this.
Next I added a media query the changes the width of the header and body of the table to be fit-content when the screen is smaller than 1477px. This was done fix an issue where the table partially disappeared.
| Small Screen | Large Screen |
| ------ | ------ |
|    ![small_screen_table](/uploads/405548632802caff2d1012c159513617/small_screen_table.png)    |    ![large_screen_table](/uploads/30a3c459d05abcaf1aacb01a5aed7732/large_screen_table.png)    |

# Note
Other reports tables will need to fixed later on.