(releases_23.3.3->develop) CLOUD-13924 - Remove the nx-button
## Problem

The new `nx-button` had some problems that were not really solvable. We will shortly in the future make a new button that follows the Material Button pattern. This allows us to not have to reimplement all of the APIs of the basic button component.

## Solution

Instead of investing more in the old button we are going to remove it. We can just use the existing global button stylings that already work fine. We will make a new button component in the Component Library coming soon

(cherry-picked from commit 2bf6067d7e83)