CLOUD-13229 - bug - Camera Status doesn't get updated in settings
## Problem

When you update the credentials on the camera settings page, sometimes the camera will still return "Unauthorized" on the first response. Sometimes the camera takes a few seconds before becoming Authorized. This causes an issue on the Camera settings page, because there is no reactivity on Camera changes there, and there is nothing to subscribe to for us to know that this value changed.

## Solution

This is the janky solution, but will 100% fix the problem and not cause other bugs. Just do a 30 second interval and check the camera's status in camera manager.


-----

### Other options

The other solutios, which if folks want we can do one of these 2 options:
1. Update all of the alerts on this page to not be set randomly throughout the component. Instead we would try to use zonejs magic and do the checks directly in the HTML by checking the `system.cameraManager.cameras[camera.id].status`.
    * `cameraManager.cameras[camera.id]` isn't a thing right now so I would need to build a map to turn the cameras array into an object as well.
    * Would need to make a decent amount of changes to the Alerts because they are all updated sporadically and manually throughout the component. Since none of these things are signals you can't just make them computeds.
2. There already is a `cameraManager.cameras$` observable for all the cameras. The problem is this does not fire an update when the status changes. So another option is to find some way to make this observable fire when the status changes. We still do our polling every 30 seconds and it does grab the update, but I guess since it's reference is not changed then it doesn't fire an update.
    * This one has more unknowns, at least for me. I'm not sure how feasible this is or how many other things would be effected if this happens.
    * Of course this also still takes some changes in the component to listen to `cameraManager.cameras$` and update the alerts accordingly

#### Pro/Cons
The pros of either of these solutions is that they are better from a high level on how we should write code. At least they are better than what I have done in this MR.

The cons are that it's possible to introduce more bugs since we will be touching more things. Even bugs outside of Cameras since lots of things interact with ``cameraManager.cameras$`. This screen will also be rewritten entirely for Cloud 2.0 so I'm not sure how much we care to invest in it still.