CLOUD-13677 - Session Truncated - Show banner when session is soon to expire
## Problem

In 6.1 we added the ability to limit the cloud session based on the `sessionLimitMinutes` in the system settings. We want to show a banner that tells the user their session will expire in X minutes if there is less than 1 hour remaining.

## Solution

The main idea is that we have a service that runs in the background, and whenever the `currentSystem` changes we setup some logic to show the banner when needed.

When the currentSystem is set, we set a timeout to start showing the banner when 1 hour remains on their session. Then every 10 seconds we check and update the time, because design wants it to be a countdown.

The Ribbon service is not great for this, because components everywhere are hiding the global ribbon constantly. I added a force show option to the ribbon service, and I know that is not a great pattern. I have an idea for how the ribbon service should be to prevent this, but with Cloud 2.0 coming and Ribbons being completely redesigned I think it's better to just force this specific use case in there.

Same idea for the way the ribbon looks. I could open up the Ribbon API and make it more robust. But this ribbon and the ribbon service will be thrown away very soon and I don't think it's worth the effort.

![image](/uploads/2695cd74e6653364a91bde3bb7edcd94/image.png)