CLOUD-14063, CLOUD-14066, CLOUD-14068: Fixes for nested partners
# Summary
Various fixes were made for supporting nested partners, routing, and the view for settings.

# Address the following tickets
CLOUD-14011, CLOUD-14036, CLOUD-14060, CLOUD-14063, CLOUD-14064, CLOUD-14066, CLOUD-14068

# CLOUD-14011
## Summary
Adjusted permission check for showing the settings tab so that normal users don't see an empty settings tab. Will need to update in the future when disconnecting from org/partner is turned on.

# CLOUD-14036
## Problem
The form updated on a delay so the description became out of sync with the dropdown.

## Solution
Wrap setting the description in a setTimeout so that the form can settle.

# CLOUD-14060
## Problem
The border line for tabs used an absolution position which broken when the height changed for nested partners.

## Solution
Make the parent of the border like relative and have the like absolute to its parent vs the page.

# CLOUD-14063, CLOUD-14064
## Problem
Permission checks were calculated in a way that resulted in the state block being hidden if the user had a role in the sub-channel.

## Solution
Checked the parent partner's permissions when showing the state block.

# CLOUD-14066
## Problem
The links to opening reports were calculated in a way that ignored sub-channel partners.

## Solution
Added the channel partner service to access all of the route ids for channel partners.

# CLOUD-14068
## Problem
If a user had lower permission in the parent partner they couldn't get to their sub-channel partner in the UI.

## Solution
Updated how the L2 header was calculated and allow nested sub-channels to show up if the parent partner had lower permissions.