CLOUD-13936: Updated logic for bind in Auth
# Problem
There's a promo page that talks about orgs. As a result we shouldn't give the user the choice between account and no orgs.

# Solution
When initializing the bind step check if the user has orgs. If they don't jump to the confirmAccount state.
Also made sure that the back button goes to the correct state from the org promo and when the user doesn't have orgs.