CLOUD-14001 - Don't create systems on the Home page
## Problem

We are currently fetching system info on the home page for every system the user has access to. With channel partners it's possible have have hundreds to even thousands of systems. In it's current state this will not scale well.

## Solution

Remove the code that is creating all of the systems on startup.

If we get lots of reports of unusable slowness we can come back and try to find solutions for this. One option is guessing/knowing which 10 systems they are most likely to access and fetching only those, either based on their user id from some backend store, or simply storing which systems they have accessed in local storage.

There is also a new feature where each system has it's own Cloud Session Limit. This causes issues since some of these systems will 404 when fetching info. In general this is fine, but we will need to address this when trying solutions.

(cherry picked from commit 8557ba092687ca8a7ecc57fc74294959728bc9ed)