CLOUD-14000 - Refactor routing for System routes
## Problem

We need to reset the `systemService.currentSystem` to undefined when we leave a system. This can allow us to listen to `currentSystem` more accurately around the app.

To do this, the easiest way is to add a `canDeactivate` on all the system routes.

## Solution

The system routes were kind of duplicated throughout instead of nested so that some of the logic could be shared. I nested these routes together and cleaned up the `canActivates`.

Then we can add a simple `canDeactivate` to reset the `currentSystem` on all the system routes.