CLOUD-14080: Fixed downloads urls
# Problem
If the app loaded on the download others page the menu didn't initialize correctly so it was broken and didn't navigate.

# Solution
Updated the paths so that they aren't broken and if platform is not set its put to a invalid string so that the router can handle it correctly.

# Note
Look into cleaning up the routing for the new downloads page.