CLOUD-14151: Refactored the month selector to support years
# Summary
When the dropdown is open it should remain open until a value is selected or the user closes it

# Note
For now we will not auto close the selector. There needs to be a more efficient way to determine it, but since its a minor function issue we leave it for now.