CLOUD-14151: Refactored the month selector to support years
# Summary
When the dropdown is open it should remain open until a value is selected or the user closes it

# Note
For now we will not auto close the selector. There needs to be a more efficient way to determine it, but since its a minor function issue we leave it for now.

(cherry picked from commit 8ae3e2d987372bc5b727a0ca3ba31b9e5f46a7ce)