CLOUD-13964, CLOUD-14173, CLOUD-14178, CLOUD-14180, CLOUD-14181: Bulk of UI partner fixes
# CLOUD-13964
## Problem
On small screens the table's scrollbar appeared which shifts the header + body.
## Solution
Added overflow-y to the reports content.

# CLOUD-14173
## Problem
Styling was applying a padding on hover that shifted the menu when hovering of items.
## Solution
Fixed the styling issue where the org left menu shifts.

# CLOUD-14178
## Problem
Partners would redirect users to Partners or Orgs where they had reports roles.
## Solution
Used filtered Orgs and Partners when setting the default redirectPath.

# CLOUD-14180
## Problem
Users are able to navigate to partners where they have the reports viewer role. They shouldn't be able to access the channel partner that way.
## Solution
Hide the go back button in the sub channel partner if the user is a reports viewer in the parent.

# CLOUD-14181
## Problem
In some weird cases users are unable to access partners where they have admin or manager due to partner nesting.
## Solution
Adjusted the filter for channel partners to show Admins and Mangers in the L2 menu.