CLOUD-14181, CLOUD-14184, CLOUD-14185, CLOUD-14186: UI fixes for partners
# CLOUD-14181
## Problem
The default channel partner is set to the last non report viewer partner.
## Solution
Updated the for loop that iterators of partners and break out of the first match.

# CLOUD-14184
| Before | After |
| ------ | ------ |
|    ![shared_before](/uploads/efd512242a59fa8adc0402a071c6d327/shared_before.png)    |    ![shared_after](/uploads/bee4257e9ea2aa9d97dbbdc9e49f2c84/shared_after.png)    |

## Problem
Filter wasn't removing owned systems in the shared tab, because the filter was moved and only applied when looking at personal systems. This was mostly caused by confusion with the variable name.
## Solution
Added the filter back onto systems and changed the filter functions name from filterByIsMine -> filterByIsOwnership.

# CLOUD-14185
## Problem
The search query param had an array with an empty string as its value ['']. This caused the search to be truthy leading to the app trying to filter by an array.
## Solution
Changed how the search$$ value was calculated. Replaced the direct usage of the router with the paramSignal

# CLOUD-14186
| Before | After |
| ------ | ------ |
|    ![folders_before](/uploads/25e0820c2995f8cdab7ef2840650001e/folders_before.png)    |    ![folders_after](/uploads/46553326c20cacdcd8c27267a36309a0/folders_after.png)    |

## Problem
Styling was slightly off from the figma for the left menu
## Changes
- Adding padding + border-radius to all elements so that each item is consistent in all states.
- Adjusted the spacer from effectively 24px to 25px.

# Other
## Placeholder alignment
| Before | After |
| ------ | ------ |
|    ![orgs_before](/uploads/172b6dd2b2bce0ccc07f9ee34af020bb/orgs_before.png)    |    ![orgs_after](/uploads/cfefcbfc2af7f871c6da8453404ca8d6/orgs_after.png)    |
|    ![info_before](/uploads/ec4773e75016860aabfa9b59fc5363b1/info_before.png)    |    ![info_after](/uploads/dece1db1524b9b307ef2faebc771852d/info_after.png)    |
### Problem
Placeholders in the organization and info tab were not lined up.
### Solution
Added a container around the action buttons to center the button.

(cherry picked from commit 00b51055d193a7ec9e27eccb4f1a7aa525e7e532)