CLOUD-14180: Adjusted routing for sub cps
# Problem
When we have nested cps (manager - manager) if the user clicks on the back button it tries to go to subchannels but fails since they can't access the tab.

# Solution
Update the CP-tab-guard so that if no route matches we redirect to the parents home page and let it route the user to a page they can view.

# Note
The default page they can view is orgs, but this is subject to change in the future depending on changes to permissions and roles.

(cherry picked from commit 86705ea21c4db8ece07c9162492ed6123498d861)