CLOUD-14201: Fixes user table not updating after folder is removed
# Problem
When a folder is deleted, and you navigate to the users tab the current group will not update so you will see all of those previous users.

# Solution
When the folder is deleted from the org force the current group in the OrgUsers Store to update.

# Changes
setSelectedGroup was updated to allow '' as a group. When passed into the method its suppose to reset the current group.

(cherry picked from commit d2c494c6dc24a1b9b04e7b7cb51cb4c8a583b513)