CLOUD-13217: Add new Service Usage Details page for expiring services
When clicking on a service in the Service Usage page, we open a new page with a detailed report for that service. We should open a different page for expiring services vs non-expiring ("regular") services. Before we only had the page for regular services; this MR adds the page for expiring services.

  

_Changes:_

- Add Expiring Service Details page + table
  - This is set up similar to the existing service usage details page
- Rename previous service usage details page + table + dialog to "regular service details" 
  - **This renaming comprises a lot of the changes since it touches many files**
- Add routes for regular vs expiring services, navigate to the right route based on service type

  

_Not added yet (part 2 of this ticket):_

- Open the expiring service details dialog when clicking on a row in the expiring service details table


Note - in the future I will break out any bulk renaming into a separate MR

![image](/uploads/5cdbf5a1db6cf64c52277cbd4341b734/image.png)