CLOUD-14219: Fixes issue with modifying org users right after being added
# Problem
The user was being updated with the wrong call due to a weird check that determines what api is used.

# Solution
Updated the check to make sure that the folder was not the org. Also had to update the users org store so that the user was updated correctly.

# Other Fixes
- Fixed issue where checkbox for all users was checked if the first row was disabled.
- Prevent extra api calls for a channel partner parent by ignore **REDACTED**
- Fixed issue in channel partner users where multiple users can be disabled when toggling roles in the partner.

(cherry picked from commit 51affda527635f106c3b1f2f3339ef039fcffa11)