CLOUD-14034: Redirect to no system access page
The no system access placeholder is supposed to be shown on the system's page, but the guard was redirecting to root instead when a user didn't have permission to access the system.

Silently redirect the user a placeholder page without modifying the url to make it behave like pre-channel partners. Also restored the no access svg accidentally removed in !7271.

![2024-07-29_16-40-12](/uploads/c65fcb7b7aba7961e5ad19671456138b/2024-07-29_16-40-12.mp4)