CLOUD-13009: Fix 5.1 Update User Issue
# Problem
When attempting to update a local user's info in System Settings, the network request was failing.

# Solution
I believe the solution has to do with how the observable was being handled. Since there were no errors showing in the network tab, I added in `pipe/catchError` to try to log the error. As soon as I added it, the error stopped happening. Removing it caused the error to happen again and adding it back in fixed it again. The retry is there for good measure.
![Screen_Recording_2024-07-25_at_9.14.56_AM](/uploads/05f9ec5ca437792ed29448ee42249d1d/Screen_Recording_2024-07-25_at_9.14.56_AM.mov)