CLOUD-13379: Fix Org Tree Icon & Padding
# Problem
When adding an org user, the folder tree was missing an icon for the top level (Org name) and the indentation of the first (and therefore subsequent) folders in the list was incorrect.

# Changes
Added icon and updated padding calculation.
![Screenshot_2024-07-31_at_3.28.05_PM](/uploads/44b3c4d8aca822e1d3766fe6a7f62872/Screenshot_2024-07-31_at_3.28.05_PM.png)

(cherry picked from commit 8a2a4a25ba3670c897839e4b063eaffb4e308bdb)