CLOUD-13417: Fix transcoding error
# Problem
Sometimes a initial mime check indicates that a stream is supported but isn't actually supported by browser which triggers the mediaserver to send another offer with transcoding.
# Solution
Mediaserver notes if an offer requires transcoding on 6.0+ added check on the signaling message. If the connection was established without transcoding being allowed then it emits an error.