CLOUD-14106: Update Access Table CSS
# Problem
For user tables (Org and CP) font weight for Administrator, Organization Administrator roles was incorrect (should be 400, not 500). The font size for the entire table was incorrect (should be 14px, not 13px). Also according to Figma (and double checked with QA), the readonly class should have the same font weight and size (400/14px).

# Changes
Updated css to match Figma
![Screenshot_2024-08-06_at_5.44.46_AM](/uploads/b9431d8f69bef78827c7752278e5e002/Screenshot_2024-08-06_at_5.44.46_AM.png)

![Screenshot_2024-08-06_at_5.44.55_AM](/uploads/20180ace451ac7873d2a234b66162be9/Screenshot_2024-08-06_at_5.44.55_AM.png)![Screenshot_2024-08-06_at_5.34.38_AM](/uploads/bb19d4464c90da3f73e73e7fb972fa41/Screenshot_2024-08-06_at_5.34.38_AM.png)

![Screenshot_2024-08-06_at_5.35.50_AM](/uploads/9e3a0c3fd18d0827119bb43b80591fae/Screenshot_2024-08-06_at_5.35.50_AM.png)![Screenshot_2024-08-06_at_5.32.51_AM](/uploads/937383bb0383529e8188690309bb9613/Screenshot_2024-08-06_at_5.32.51_AM.png)