CLOUD-14249  - Cloud Session Limiter - Move Toggle into main settings section
## Problem

To turn on the Cloud Session Limiter, previously you had to go to advanced settings. Product wants the settings visible and next to the existing Session Limit duration. Design also wants the whole section to be hidden when Session Limit duration is turned off since there is an alert block there.

## Solution

Expose the setting value and add the inputs to the form. Then hide the whole section if the main Session Limit is turned off.

| BEFORE | AFTER |
| ------ | ------ |
|    ![image](/uploads/f8593ba5bccc03fcae656b35ec4c4106/image.png)    |    ![image](/uploads/3e315a7311d425fc1d5b3e99286bdb1c/image.png)    |
|   ![image](/uploads/a0bb0a58bc37773db12cf23a638f9444/image.png)     |    ![image](/uploads/ba9a0cda6334ba0855911c16ea62f0f8/image.png)    |