CLOUD-14254, CLOUD-14255: System User fixes
# Problem
Users were processed in a way that made it so that
- Power users could edit each other (All 6.0 systems)
- A channel partner Administrator couldn't edit power users (Any 6.0 org system)

# Solution
Updated how users are processed so that a channel partner user can be set as the current user allowing them to edit power users that were added to the system. To fix power users editing each other isCloudOwner needed to be updated to not check `isAdmin`. This effectively made them similar to the owner allowing them to edit other power users. Note: It was a UI bug so the server was fine, and pre 6.0 systems the logic was good.

# Other Changes
- Refactored process users in both manager classes so that process user can be used independently of process users. This will be helpful later for improving jsonrpc.

(cherry picked from commit df4881b7c72d80b3c18c336941f885c4dcdcc09b)