CLOUD-13999: Hide names of indirect descendants in reports
In channel partner reports, a user should only be able to see names of child partners/orgs that are direct children (one level deep), unless they are also a member of the nested partner/org.

This is implemented in the API which returns `**REDACTED**` for the names of these entities. This MR updates the frontend to show a `Hidden Name` link which copies to clipboard the url to that entity's service report, per the [spec](https://networkoptix.atlassian.net/wiki/spaces/FS/pages/2996142086/SaaS+Reports#I.3.-General.-Remark-About-Hidden-Names).

![image](/uploads/362d3cb2042307cbd47c13f5e288f8c7/image.png){width=80%}
![image](/uploads/beb65ba7735825d5ce111e0f0309a80e/image.png){width=40%}
![image](/uploads/5d6c086c94bde13f270f504b653fe0bd/image.png){width=80%}
![image](/uploads/933bbd94e2cea54006ce0c7264a3808b/image.png){width=80%}