CLOUD-14415, CLOUD-14417: Fixed session refresh logic for changing passwords
# Problem
The process for renewing a users session for changing the password was hardcoded to admin. As a result non admin users couldn't change there password or if their current password matched the admin the session would change to the admin account.

# Solution
Remove the hardcoded 'admin' and use the selected user's login name. If the user is admin nothing changes, and now non-admin users can change their password.