CLOUD-14447 - View Tab has no servers or cameras in list
## Problem

In a previous change (https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/8110/diffs?view=parallel#49fee8723b00b8bc7fda4c4682fae836b8f74e41_16_13) the `nx-media-server-list` was removed from the sidebar. I think this was done by accident so I'm adding it back to how it was before the change.

## Solution

The whole section inside the `@else` I just copy and pasted back in so it is how it was before the change