CLOUD-14386: Fix error handling in regular + expiring service pages
When a service report is not available, the API sends a response with the format `{ detail: errorMessage }`. The service usage page was set up to handle this, but regular/expiring service pages were not - this MR updates them to match. Also clear any existing error on load for the expiring service page.

![image](/uploads/1f63a867421b533bdc7d639ea974d4ee/image.png)