CLOUD-14490 - View Tab breaks on refresh
## Problem

Something a while ago broke the view tab so that you can't load into the page directly. It throws an error because the `currentSystemResolver` throws an error for the `systemManager.currentSystem` being undefined.

## Solution

To fix this in Cloud Portal we use the `SystemGuard`. There is probably a solution where we don't use this, but this has been blocking QA for a long time so we need to get some fix in. And since this has been broken for so long I wasn't able to find which commit broke it to try and fix it some other way.