CLOUD-14049 CLOUD-13893 Updating dialog stylings for bookmark sharing and new filters
## Problem

There was some updates to how dialogs are styled, which allows us to make use of shared styling. This simplifies and centralizes things nicely, but in the meantime it broke some existing dialogs that were using the DialogV2Factory.

## Solution

Migrate these 2-ish dialogs over to use the new styling format

| BEFORE | AFTER |
| ------ | ------ |
|    ![image](/uploads/a5cd933aefe5b7349469915256adb4a3/image.png)    |    ![image](/uploads/7d41e6f293f6d5e57438852c52133fef/image.png)    |
|   ![image](/uploads/0b8b01f8cf264f83a9ae716c8c332318/image.png)     |     ![image](/uploads/8fc00cf26e8511e44877f563b75ff3c4/image.png)   |