CLOUD-14312 CLOUD-14313 - fix retry on get for legacy API
## Problem

There were timing issues when trying to load Cameras and Users too quickly. This is because the `Auth` param was not added to the request. This is actually fine because we retry the request and by then the auth param is ready.

BUT there is an existing bug that the `retryHandler` doesn't actually get the updated `Auth` param and add it to the request before trying again.

## Solution

There isn't a super easy way to alter the request from inside the `retryWhen`. Instead of refactoring the whole error handling process, I just altered the flow of the rxjs so that it gets a new request each time the error happens.