CLOUD-14476: Reset startTs to all services value
# Problem
When a user goes to the detailed service view they can change the date and when they go back to all services it will be reporting on a different time from when they started. It may cause confusion when the report is different.

# Solution
Save the initial startTs when they navigate to the detailed view. When the user tries to go back change the startTs value to the initial value when they went to the detailed view.

# Gotcha
If the user reloads on the detailed view they will go back to all services with that value. This is a limitation from how the date saving was implemented.

(cherry picked from commit 18831e757c17f528653a752821390d29fdcd00b4)