CLOUD-14278: Fix broken Bookmark thumbnail for zoomed in screens
# Problem

When the screen is zoomed in on the Bookmarks page, the Thumbnails and videos sometimes breaks. We calculate the size of the height/width of the thumbnails like so:

* Width: 270 x aspectRatio x `dpr`
* Height: 270 x `dpr`

`dpr` = Device Pixel ratio

When we zoom in, sometimes the `dpr` will give us a decimal number, this is what's breaking our thumbnail. If the thumbnail is broken, we don't display the video for the user

**Normal screen size API call:**

```typescript
62205d7c-4da2-417c-b1a3-1f838a00bb19.relay.relay.cloud.hdw.mx/ec2/cameraThumbnail?cameraId=c444a370-b07f-80b5-c26c-33ceb1a70c5f&time=1723646685265&streamSelectionMode=forcedPrimary&width=480&height=270
```

**Zoomed in 125% API call:**

```typescript
62205d7c-4da2-417c-b1a3-1f838a00bb19.relay.relay.cloud.hdw.mx/ec2/cameraThumbnail?cameraId=c444a370-b07f-80b5-c26c-33ceb1a70c5f&time=1723646685265&streamSelectionMode=forcedPrimary&width=600&height=337.5
```

* Notice how at the end of the API call, the height is `337.5`

# Changes

Round up the thumbnail's height/width to the next integer

Zoomed in 125%:

![Screenshot 2024-08-21 at 5.05.57 PM.png](/uploads/e5a7d2c6f8076c0545ba2bef0972c7cb/Screenshot_2024-08-21_at_5.05.57_PM.png)

(cherry picked from commit 2ba0d189a12aa450b9b02707948512421c24bdf8)

# Conflicts:
#	front_end/libs/features/systems/bookmarks/bookmarks.component.ts