CLOUD-12400 - Only edit Motion Mask when user can view AND edit camera
## Problem

Design wants us to disable to Motion Mask settings if a user does not have View AND Edit permissions for a camera. Technically the user has permissions to edit the motion mask, but since they can't see what they are editing we should just disable it for now.

## Solution

Add another conditional to when we should enable the overlay. This time only if the user can view the device.

Maybe there is a world where we should show the motion mask overlay on top of the "No Image" placeholder and allow the user to edit it since technically they have permissions. But the current motion mask code breaks if you try to show "No Image" placeholder and I didn't immediately know how to fix it. Seems like a really niche issue anyway.

Duplicate MR from https://gitlab.nxvms.dev/dev/cloud_portal/-/merge_requests/8369